checkpoint

+++ reel.h
@@ -5,13 +5,16 @@
 #include "rationnel.h"
 
 class Reel: public Constante{
-    float _valeur;
+    //float _valeur;
+    double _valeur;
 
 public:
-    Reel(float e):_valeur(e){}
+    ///Reel(float e):_valeur(e){}
+    Reel(double e):_valeur(e){}
     Reel(Constante* c);
     virtual QString afficher() const{ return QString::number(_valeur);}
-    float getValeur() const {return _valeur;}
+    //float getValeur() const {return _valeur;}
+    double getValeur() const {return _valeur;}
     void setValeur(int v){_valeur=v;}
 
     Rationnel toRationnel() const;
